client - completed entire post-details functionality

+++ api/functions/requests/comment/like_comment.js
@@ -66,29 +66,31 @@ exports.likeComment = async (req, res) => {
                 occupation: comment_data.owner.occupation,
             }).catch(() => { throw Error('There was an error deleting your post. Please try again.') });
             
-            await createNewNotification({ 
-                batch, 
-                timestamp, 
-                local_uid, 
-                notification_type: 'comment_liked',
-                notification_owner_uid: comment_data?.owner.uid, 
-                content: { 
-                    ref_id: comment_id, 
-                    post_ref_id: post_id, 
-                    text: comment_data.text,
-                },
-            }).catch(() => { throw Error('There was an error deleting your post. Please try again.') });
+            if(comment_data.owner.uid !== local_uid) {
+                await createNewNotification({ 
+                    batch, 
+                    timestamp, 
+                    local_uid, 
+                    notification_type: 'comment_liked',
+                    notification_owner_uid: comment_data?.owner.uid, 
+                    content: { 
+                        ref_id: comment_id, 
+                        post_ref_id: post_id, 
+                        text: comment_data.text,
+                    },
+                }).catch(() => { throw Error('There was an error deleting your post. Please try again.') });
+            };
 
             await batch.commit()
                 .catch(() => { throw Error('An internal error occurred. Please try again') });
 
-            res.status(200).send({ message: 'Comment liked!' });
+            res.status(200).send({ message: 'Comment liked!', comment_liked: true });
         }
 
         if(!comment_liked) {
             const post_data = (await post.get()).data();
             const batch = getFirestore().batch();
-
+            
             await getFirestore().collection(`comments/${post_id}/liked_comments`)
                 .where('comment_id_ref', '==', comment_id)
                 .where('owner.uid', '==', local_uid)
@@ -98,25 +100,32 @@ exports.likeComment = async (req, res) => {
 
             batch.set(comment, { total_likes: FieldValue.increment(-1) }, { merge: true });
 
-            await userActivityHistory({ local_uid, batch, type: 'comment_like', other_user_uid: post_owner.owner.uid, post_id, comment_id })
-                .catch(() => { throw Error('An internal error occurred. Please try again') });
+            await userActivityHistory({ 
+                batch, local_uid, 
+                type: 'comment_like', 
+                post_id,
+                comment_id,
+                other_user_uid: comment_data?.owner.uid, 
+            }).catch(() => { throw Error('An internal error occurred. Please try again') });
             
-            await deleteNotificationEntry({ 
-                batch, 
-                timestamp, 
-                local_uid, 
-                notification_type: 'comment_liked',
-                notification_owner_uid: comment_data?.owner.uid, 
-                content: { 
-                    ref_id: comment_id, 
-                    post_ref_id: post_id, 
-                },
-            }).catch(() => { throw Error('There was an error deleting your post. Please try again.') });
+            if(comment_data.owner.uid !== local_uid) {
+                await deleteNotificationEntry({ 
+                    batch, 
+                    timestamp, 
+                    local_uid, 
+                    notification_type: 'comment_liked',
+                    notification_owner_uid: comment_data?.owner.uid, 
+                    content: { 
+                        ref_id: comment_id, 
+                        post_ref_id: post_id, 
+                    },
+                }).catch(() => { throw Error('There was an error deleting your post. Please try again.') });
+            };
 
             await batch.commit()
                 .catch(() => { throw Error('An internal error occurred. Please try again') });
 
-            res.status(200).send({ message: 'Comment unliked!' });
+            res.status(200).send({ message: 'Comment unliked!', comment_liked: false });
         }
 
     } catch(error) {
